improvement(react-cms): move function to useSource
